refactor(studyset): change the study set responses and flashcard

+++ Flashcards/FlashcardService.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
+using FlashcardXpApi.Auth;
 using FlashcardXpApi.Common.Results;
 using FlashcardXpApi.Flashcards.Requests;
 using FlashcardXpApi.FlashcardSets;
+using FlashcardXpApi.Users;
 
 namespace FlashcardXpApi.Flashcards
 {
@@ -12,13 +14,15 @@ namespace FlashcardXpApi.Flashcards
         private readonly ILogger _logger;
         private readonly IMapper _mapper;
         private readonly FlashcardRequestValidator _validator;
+        private readonly AuthService _authService;
 
-        public FlashcardService(
+        public FlashcardService (
             IFlashcardRepository flashcardRepo, 
             IStudySetRepository studySetRepo, 
             ILogger<FlashcardService> logger, 
             IMapper mapper, 
-            FlashcardRequestValidator validator
+            FlashcardRequestValidator validator,
+            AuthService authService
         )
         {
             _flashcardRepo = flashcardRepo;
@@ -26,23 +30,36 @@ namespace FlashcardXpApi.Flashcards
             _logger = logger;
             _mapper = mapper;
             _validator = validator;
+            _authService = authService;
         }
 
-        /*
-        public async Task<ResultGeneric<List<FlashcardDto>>> GetAllByStudySet(
+        
+        public async Task<ResultGeneric<FlashcardsByStudySetDto>> GetAllByStudySet(
             string studySetId
         )
         {
+            var user = await _authService.GetLoggedInUser();
+
             var studySet = await _studySetRepo.GetByIdAsync(studySetId);
 
             if (studySet is null)
             {
-                return ResultGeneric<List<FlashcardDto>>.Failure(FlashcardErrors.StudySetNotFoundError);
+                return ResultGeneric<FlashcardsByStudySetDto>.Failure(FlashcardErrors.StudySetNotFoundError);
             }
-
+            
             var flashcards = await _flashcardRepo.GetAllByStudySet(studySet);
-            return ResultGeneric<List<FlashcardDto>>.Success(
+
+            var response = new FlashcardsByStudySetDto
+            (
+                studySet.Id,
+                studySet.Title,
+                studySet.Description,
+                _mapper.Map<UserDto>(studySet.CreatedBy),
                 _mapper.Map<List<FlashcardDto>>(flashcards)
+            );
+            
+            return ResultGeneric<FlashcardsByStudySetDto>.Success(
+                response
             );
         }
 
@@ -71,10 +88,15 @@ namespace FlashcardXpApi.Flashcards
                 return ResultGeneric<List<FlashcardDto>>.Failure(FlashcardErrors.StudySetNotFoundError);
             }
 
+            var user = await _authService.GetLoggedInUser();
+
+            if (user is not null && studySet.CreatedById != user.Id)
+            {
+                return ResultGeneric<List<FlashcardDto>>.Failure(FlashcardErrors.NotAuthorizedError);
+            }
 
             var newFlashcards = _mapper.Map<List<Flashcard>>(requests);
             newFlashcards.ForEach(f => f.StudySet = studySet);
-
 
             _logger.LogInformation("Adding new flashcard");
             await _flashcardRepo.InsertAllAsync(newFlashcards);
@@ -96,6 +118,13 @@ namespace FlashcardXpApi.Flashcards
                 return ResultGeneric<FlashcardDto>.Failure(FlashcardErrors.FlashcardNotFoundError);
             }
 
+            var user = await _authService.GetLoggedInUser();
+
+            if (user is not null && flashcard.StudySet.CreatedById != user.Id)
+            {
+                return ResultGeneric<FlashcardDto>.Failure(FlashcardErrors.NotAuthorizedError);
+            }
+
             await _flashcardRepo.DeleteAsync(flashcard);
 
             return ResultGeneric<FlashcardDto>.Success (
@@ -103,6 +132,5 @@ namespace FlashcardXpApi.Flashcards
             );
         }
         
-        */
     }
 }
